fix value underflow
In the default case the ISAAC index type is unsigned. During the mapping
from the value to the transfere function index type a value underflow
and happened and therefore negative values get mapped to the max
transfere function value. The result are artefacts during the
visualization.

Fix this behaviour by clamping with float values before we cast back to
an index type.